requirements.txt
openai==1.54.4
pdf2image==1.17.0
Pillow==11.0.0
pytesseract==0.3.13
streamlit==1.40.1
toml==0.10.2
